+++ NEWS.md
@@ -9,7 +9,7 @@
 ## Enhancements
 
 * Show an empty progress bar just before targets start to build when `verbose` is `2` (#1203, @kendonB).
-
+* Deprecate the `jobs` argument of `clean()`.
 
 # Version 7.11.0